Adapt form.selected_fitlers() tests to view tests
