+++ etna/search/tests/test_forms.py
@@ -32,112 +32,3 @@ class CatalogueSearchFormTest(SimpleTestCase):
         is_valid = form.is_valid()
 
         self.assertTrue(is_valid)
-
-
-class CataglogueSearchFormSelectedFiltersTest(SimpleTestCase):
-    def test_with_valid_filter_values(self):
-        form = CatalogueSearchForm(
-            {
-                "group": "tna",
-                "topic": ["topic-one"],
-                "level": ["Division"],
-                "collection": ["WO", "AK"],
-            }
-        )
-
-        self.assertTrue(form.is_valid())
-
-        self.assertEqual(
-            form.selected_filters,
-            {
-                "collection": [
-                    (
-                        "WO",
-                        "WO - War Office, Armed Forces, Judge Advocate General, and related bodies",
-                    ),
-                    ("AK", "AK - County Courts"),
-                ],
-                "level": [("Division", "Division")],
-                "topic": [("topic-one", "topic-one")],
-            },
-        )
-
-    def test_counts_are_removed_from_updated_choice_labels(self):
-        form = CatalogueSearchForm(
-            {
-                "group": "tna",
-                "topic": ["topic-one"],
-                "level": ["Division"],
-            }
-        )
-        self.assertTrue(form.is_valid())
-
-        topic_field = form.fields["topic"]
-        level_field = form.fields["level"]
-
-        # Update topic field choices to include counts on labels
-        topic_field.update_choices(
-            [
-                {"key": "topic-one", "doc_count": 10},
-            ]
-        )
-        self.assertEqual(topic_field.choices, [("topic-one", "topic-one (10)")])
-
-        # Update level field choices to include counts on labels
-        level_field.update_choices(
-            [
-                {
-                    "key": "Division",
-                    "doc_count": 10,
-                },
-            ]
-        )
-        self.assertEqual(level_field.choices, [("Division", "Division (10)")])
-
-        self.assertEqual(
-            form.selected_filters,
-            {
-                "topic": [("topic-one", "topic-one")],
-                "level": [("Division", "Division")],
-            },
-        )
-
-    def test_with_invalid_filter_values(self):
-        form = CatalogueSearchForm(
-            {
-                "group": "tna",
-                # only the 'level' field validates values against the choices
-                "level": ["foo"],
-            }
-        )
-        self.assertFalse(form.is_valid())
-        self.assertEqual(form.selected_filters, {})
-
-    def test_with_partially_invalid_filter_values(self):
-        form = CatalogueSearchForm(
-            {
-                "group": "tna",
-                "topic": ["topic-one", "topic-two"],  # valid
-                "catalogue_source": ["catalogue-source-one"],  # valid
-                "collection": ["bar"],
-                "level": ["foo"],  # invalid
-            }
-        )
-
-        self.assertFalse(form.is_valid())
-
-        self.assertEqual(
-            form.selected_filters,
-            {
-                "collection": [
-                    ("bar", "bar"),
-                ],
-                "topic": [
-                    ("topic-one", "topic-one"),
-                    ("topic-two", "topic-two"),
-                ],
-                "catalogue_source": [
-                    ("catalogue-source-one", "catalogue-source-one"),
-                ],
-            },
-        )
