+++ etna/search/tests/test_views.py
@@ -1,3 +1,4 @@
+from typing import Any
 from unittest import mock
 
 from django.http import HttpResponse
@@ -7,6 +8,120 @@ from django.urls import reverse
 import responses
 
 from ...ciim.tests.factories import create_response
+from ..forms import CatalogueSearchForm
+from ..views import CatalogueSearchView
+
+
+class SelectedFiltersTest(SimpleTestCase):
+    def get_result(self, form):
+        return CatalogueSearchView().get_selected_filters(form)
+
+    def test_with_valid_filter_values(self):
+        form = CatalogueSearchForm(
+            {
+                "group": "tna",
+                "topic": ["topic-one"],
+                "level": ["Division"],
+                "collection": ["WO", "AK"],
+            }
+        )
+
+        self.assertTrue(form.is_valid())
+
+        self.assertEqual(
+            self.get_result(form),
+            {
+                "collection": [
+                    (
+                        "WO",
+                        "WO - War Office, Armed Forces, Judge Advocate General, and related bodies",
+                    ),
+                    ("AK", "AK - County Courts"),
+                ],
+                "level": [("Division", "Division")],
+                "topic": [("topic-one", "topic-one")],
+            },
+        )
+
+    def test_counts_are_removed_from_updated_choice_labels(self):
+        form = CatalogueSearchForm(
+            {
+                "group": "tna",
+                "topic": ["topic-one"],
+                "level": ["Division"],
+            }
+        )
+        self.assertTrue(form.is_valid())
+
+        topic_field = form.fields["topic"]
+        level_field = form.fields["level"]
+
+        # Update topic field choices to include counts on labels
+        topic_field.update_choices(
+            [
+                {"key": "topic-one", "doc_count": 10},
+            ]
+        )
+        self.assertEqual(topic_field.choices, [("topic-one", "topic-one (10)")])
+
+        # Update level field choices to include counts on labels
+        level_field.update_choices(
+            [
+                {
+                    "key": "Division",
+                    "doc_count": 10,
+                },
+            ]
+        )
+        self.assertEqual(level_field.choices, [("Division", "Division (10)")])
+
+        self.assertEqual(
+            self.get_result(form),
+            {
+                "topic": [("topic-one", "topic-one")],
+                "level": [("Division", "Division")],
+            },
+        )
+
+    def test_with_invalid_filter_values(self):
+        form = CatalogueSearchForm(
+            {
+                "group": "tna",
+                # only the 'level' field validates values against the choices
+                "level": ["foo"],
+            }
+        )
+        self.assertFalse(form.is_valid())
+        self.assertEqual(self.get_result(form), {})
+
+    def test_with_partially_invalid_filter_values(self):
+        form = CatalogueSearchForm(
+            {
+                "group": "tna",
+                "topic": ["topic-one", "topic-two"],  # valid
+                "catalogue_source": ["catalogue-source-one"],  # valid
+                "collection": ["bar"],
+                "level": ["foo"],  # invalid
+            }
+        )
+
+        self.assertFalse(form.is_valid())
+
+        self.assertEqual(
+            self.get_result(form),
+            {
+                "collection": [
+                    ("bar", "bar"),
+                ],
+                "topic": [
+                    ("topic-one", "topic-one"),
+                    ("topic-two", "topic-two"),
+                ],
+                "catalogue_source": [
+                    ("catalogue-source-one", "catalogue-source-one"),
+                ],
+            },
+        )
 
 
 class SearchViewTestCase(SimpleTestCase):
